Commit dist/ directory so this fork works with Webpack 6
This package's package.json references dist/ files (as well as .mjs and .es6.js files). Normally those would be created as part of a build process. But we're pulling from Github, so they need to be committed to be present.

Without this, our Webpack 6 setup blows up because it can't find the "browser" file, since it's not present.