Commit dist/ directory and .mjs file (#19)
* Commit dist/ directory so this fork works with Webpack 6

This package's package.json references dist/ files (as well as .mjs and .es6.js files). Normally those would be created as part of a build process. But we're pulling from Github, so they need to be committed to be present.

Without this, our Webpack 6 setup blows up because it can't find the "browser" file, since it's not present.

* Revert "Commit dist/ directory so this fork works with Webpack 6"

This reverts commit ed8cf2d3d3cde7433cad456def490809db35c4d8.

* Point bundlers to the "main" entry, instead of "browser"

To work with Webpack 6. Otherwise we would need to ensure the dist/ and .mjs and .es6.js files actually exist and are committed to the repo.